Basic documentation theme
This is largely based on how Bootstrap render documentation [1] as it
makes for a quite pleasant reading experience.

The ToC is generated client-side in Javascript which makes me a little
sad, but presuming that this will be hosted on GitHub we are limited in
plugin choice or arbitrary-code execution which is problematic. I could
dodge this and use the built-in `kramdown` functionality, but then we'd
lose the ability to work from seperate files.

The bootstrap is a stripped-down version that gives us what we need to
get a pleasant experience.

[1]: e.g. http://getbootstrap.com/css/