Add first pass regarding course information
This is based on some of my earlier work in suggesting how courses could
be described better to appeal to an international audience. Largely,
this is focused on adding elements that align our content closer to HE
courses in general.

Much of this content doesn't exist yet for courses, and some of it may
be better placed as metadata rather than copy (e.g. learning outcomes)