Introduce core OC concepts
This commit adds a section detailing the key "objects" in
OpenClassrooms' world. Currently, there are areas where we are less
consistent in our language (around exercises/assignments/assessments
etc. for instance) which creates ambiguity and confusion. A controlled
vocabulary like this provides a point of reference to ensure that we are
talking consistently across the product and courses.